# docs/technical/api/openapi.yaml
openapi: 3.0.3
info:
  title: CLARIHR APIs
  version: 0.0-bootstrap
  description: |
    Partial hand-maintained OpenAPI supplement for externally relevant core and backoffice
    endpoints that need a checked-in contract before automated export/versioning is introduced.

    The exhaustive contract is still generated at runtime by Swagger in Development
    at /swagger/v1/swagger.json.

    Public API contracts expose `publicId` or `<entity>PublicId`; internal `id` and
    `internalId` never leave the backend. When a resource has business codes, the
    public contract exposes `code` and `normalizedCode`, both normalized to uppercase.

    Core API endpoints require tokens emitted with `client_type=core`. Platform
    backoffice endpoints under `/api/platform/*` require tokens emitted with
    `client_type=platform`.

    All `/api` routes emit `Cache-Control: no-store`, `Pragma: no-cache`, and
    `Expires: 0` to avoid caching sensitive HR or platform responses.
servers:
  - url: /
security:
  - Bearer: []
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CompanyUserRoleResponse:
      type: object
      required:
        - id
        - name
        - isSystemRole
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        isSystemRole:
          type: boolean
    UserStatus:
      type: string
      enum:
        - PendingActivation
        - Active
        - Inactive
    CompanyUserResponse:
      type: object
      required:
        - id
        - roles
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        roles:
          type: array
          items:
            $ref: '#/components/schemas/CompanyUserRoleResponse'
        status:
          allOf:
            - $ref: '#/components/schemas/UserStatus'
          nullable: true
...
    PersonnelFileRecordType:
      type: string
      enum:
        - Candidate
        - Employee
    PersonnelFileLifecycleStatus:
      type: string
      enum:
        - Draft
        - Completed
...
    PersonnelFileContractResponse:
      type: object
      description: Partial checked-in contract for personnel file create, personal-info update and finalize flows.
      required:
        - id
        - companyId
        - recordType
        - lifecycleStatus
        - firstName
        - lastName
        - fullName
        - birthDate
        - age
        - isActive
        - concurrencyToken
        - createdAtUtc
      properties:
        id:
          type: string
          format: uuid
        companyId:
          type: string
          format: uuid
        recordType:
          $ref: '#/components/schemas/PersonnelFileRecordType'
        lifecycleStatus:
          $ref: '#/components/schemas/PersonnelFileLifecycleStatus'
        firstName:
          type: string
        lastName:
          type: string
        fullName:
          type: string
        birthDate:
          type: string
          format: date-time
        age:
          type: integer
          format: int32
        maritalStatusCode:
          type: string
          nullable: true
        maritalStatusName:
          type: string
          nullable: true
        professionCode:
          type: string
          nullable: true
        professionName:
          type: string
          nullable: true
        birthCountryCode:
          type: string
          nullable: true
        birthCountryName:
          type: string
          nullable: true
        birthDepartmentCode:
          type: string
          nullable: true
        birthDepartmentName:
          type: string
          nullable: true
        birthMunicipalityCode:
          type: string
          nullable: true
        birthMunicipalityName:
          type: string
          nullable: true
        institutionalEmail:
          type: string
          format: email
          nullable: true
        orgUnitId:
          type: string
          format: uuid
          nullable: true
        assignedPositionSlotId:
          type: string
          format: uuid
          nullable: true
        linkedUserId:
          type: string
          format: uuid
          nullable: true
        isActive:
          type: boolean
        concurrencyToken:
          type: string
          format: uuid
        createdAtUtc:
          type: string
          format: date-time
        modifiedAtUtc:
          type: string
          format: date-time
          nullable: true
        identifications:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              identificationTypeCode:
                type: string
              identificationTypeName:
                type: string
                nullable: true
              identificationNumber:
                type: string
              issuedDate:
                type: string
                format: date-time
                nullable: true
              expiryDate:
                type: string
                format: date-time
                nullable: true
              issuer:
                type: string
                nullable: true
              isPrimary:
                type: boolean
...
    FinalizePersonnelFileRequest:
      type: object
      required:
        - concurrencyToken
      properties:
        concurrencyToken:
          type: string
          format: uuid
        createUserAccount:
          type: boolean
          nullable: true
          description: When omitted, defaults to true. Set false to complete the personnel file without auto-provisioning a company user.
    FinalizePersonnelFilePreviewIssueResponse:
      type: object
      required:
        - code
        - message
        - section
        - fieldKey
        - navigationKey
        - isBlocking
      properties:
        code:
          type: string
        message:
          type: string
        section:
          type: string
        fieldKey:
          type: string
        navigationKey:
          type: string
          description: UI navigation hint for frontend redirection (for example `personnel-files`, `personal-info`, `employee-profile`).
        isBlocking:
          type: boolean
    FinalizePersonnelFilePreviewResponse:
      type: object
      required:
        - personnelFileId
        - createUserAccount
        - isEligible
        - issues
      properties:
        personnelFileId:
          type: string
          format: uuid
        createUserAccount:
          type: boolean
        isEligible:
          type: boolean
        issues:
          type: array
          items:
            $ref: '#/components/schemas/FinalizePersonnelFilePreviewIssueResponse'
    FinalizePersonnelFileResponse:
      type: object
      required:
        - personnelFile
      properties:
        personnelFile:
          $ref: '#/components/schemas/PersonnelFileContractResponse'
        user:
          allOf:
            - $ref: '#/components/schemas/CompanyUserResponse'
          nullable: true
        invitationExpiresUtc:
          type: string
          format: date-time
          nullable: true
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
        code:
          type: string
          nullable: true
        traceId:
          type: string
          nullable: true
          description: Correlation identifier that links the error response with server logs.
        errors:
          type: object
          nullable: true
          description: Field-level validation messages. Present when `code` is `common.validation`.
          additionalProperties:
            type: array
            items:
              type: string
        details:
          type: array
          nullable: true
          description: Optional structured metadata returned by some authorization or policy failures.
          items:
            $ref: '#/components/schemas/ProblemDetailItem'
    ProblemDetailItem:
      type: object
      properties:
        resourceKey:
          type: string
          nullable: true
        action:
          type: string
          nullable: true
        fieldKey:
          type: string
          nullable: true
        endpoint:
          type: string
          nullable: true
...
paths:
...
  /api/v1/personnel-files/{id}/finalize:
    post:
      tags:
        - personnel-files
      operationId: FinalizePersonnelFile
      summary: Complete an employee personnel file with optional user provisioning
      description: Completes a draft employee personnel file. When createUserAccount is true (or omitted), validates the configured slot role, creates or reuses the company user and issues the temporary activation invitation; when false, completes without provisioning a user.
      security:
        - Bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinalizePersonnelFileRequest'
      responses:
        '200':
          description: Personnel file completed successfully. Company-user provisioning data is returned only when createUserAccount is true (or omitted).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinalizePersonnelFileResponse'
        '400':
          description: Validation error, such as missing concurrency token.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Authentication is required.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: The authenticated user does not have permission to manage personnel files or the tenant scope does not match.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: The personnel file or its assigned position slot was not found in the tenant scope.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Business conflict such as stale concurrency token or an institutional email already linked to another personnel file.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Business-rule violation, such as trying to finalize a non-employee file, missing institutional email, missing slot, or missing slot role when createUserAccount is true.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/personnel-files/{id}/finalize/preview:
    get:
      tags:
        - personnel-files
      operationId: PreviewFinalizePersonnelFile
      summary: Validate if an employee personnel file is ready to finalize
      description: Evaluates the same functional rules used by finalize and returns eligibility plus blocking issues for frontend pre-check workflows.
      security:
        - Bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: createUserAccount
          in: query
          required: false
          description: When omitted, defaults to true. Set false to skip account-provisioning checks in preview.
          schema:
            type: boolean
      responses:
        '200':
          description: Finalization readiness evaluated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinalizePersonnelFilePreviewResponse'
        '400':
          description: Validation error, such as invalid query parameter format.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Authentication is required.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: The authenticated user does not have permission to manage personnel files or the tenant scope does not match.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: The personnel file was not found in the tenant scope.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...